+++ e2e/testdata/alter/add_column.sql
