+++ e2e/testdata/alter/drop_column.sql
