+++ e2e/testdata/alter/drop_default.sql
