+++ e2e/testdata/alter/drop_not_null.sql
