+++ e2e/testdata/alter/edit_type.sql
