+++ e2e/testdata/create_index/partial_index.sql
