+++ e2e/testdata/create_index/using_index.sql
