+++ e2e/testdata/create_table/create.sql
