+++ e2e/testdata/drop_index/drop.sql
