+++ e2e/testdata/drop_index/single.sql
