+++ e2e/testdata/drop_table/drop_table.sql
