+++ e2e/testdata/select/group_by_join.sql
