+++ e2e/testdata/select/inner_join.sql
