+++ e2e/testdata/select/lateral_join.sql
