+++ e2e/testdata/select/left_join.sql
