+++ e2e/testdata/select/moving_average.sql
