+++ e2e/testdata/select/right_join.sql
