+++ e2e/testdata/select/running_sum.sql
